add more endpoints for advertiser; refactor a field in UserBank table

+++ SEP490.AffiliateNetwork/ANF.Core/Models/Requests/UserBankCreateRequest.cs
@@ -5,7 +5,7 @@ namespace ANF.Core.Models.Requests
     public class UserBankCreateRequest
     {
         [Required(ErrorMessage = "Banking number is required!")]
-        public int BankingNo { get; set; }
+        public long BankingNo { get; set; }
 
         [Required(ErrorMessage = "Banking provider is required!")]
         public string BankingProvider { get; set; } = null!;
